refactor: 상수 모음인 LottoPolicy 클래스를 LottoConstant로 변경
