+++ src/main/java/lotto/constant/LottoConstant.java
@@ -1,8 +1,10 @@
 package lotto.constant;
 
-public final class LottoPolicy {
+public final class LottoConstant {
     public static final int LOTTO_PRICE = 1_000;
 
     public static final int MIN_PURCHASE_AMOUNT = 1_000;
     public static final int MAX_PURCHASE_AMOUNT = 100_000_000;
+
+    private LottoConstant() {}
 }
