+++ src/main/java/lotto/domain/PurchaseAmount.java
@@ -1,11 +1,11 @@
 package lotto.domain;
 
-import static lotto.constant.LottoPolicy.MAX_PURCHASE_AMOUNT;
-import static lotto.constant.LottoPolicy.MIN_PURCHASE_AMOUNT;
+import static lotto.constant.LottoConstant.MAX_PURCHASE_AMOUNT;
+import static lotto.constant.LottoConstant.MIN_PURCHASE_AMOUNT;
 import static lotto.enums.ErrorMessage.INVALID_PURCHASE_AMOUNT_RANGE;
 import static lotto.enums.ErrorMessage.INVALID_PURCHASE_AMOUNT_UNIT;
 
-import lotto.constant.LottoPolicy;
+import lotto.constant.LottoConstant;
 
 public class PurchaseAmount {
     private final int purchaseAmount;
@@ -22,7 +22,7 @@ public class PurchaseAmount {
                     INVALID_PURCHASE_AMOUNT_RANGE.getMessage()
             );
         }
-        if (amount % LottoPolicy.LOTTO_PRICE != 0) {
+        if (amount % LottoConstant.LOTTO_PRICE != 0) {
             throw new IllegalArgumentException(
                     INVALID_PURCHASE_AMOUNT_UNIT.getMessage()
             );
